requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bitarray==3.0.0
bleach==6.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
comm==0.2.2
cytoolz==1.0.0
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==1.0.1
eth-typing==5.0.1
eth-utils==2.3.2
executing==2.1.0
fastjsonschema==2.21.1
fqdn==1.5.1
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.7
httpx==0.28.0
hyperliquid-python-sdk==0.9.0
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.4
jiter==0.8.0
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.2.6
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
markupsafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.0.2
msgpack==1.1.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook-shim==0.2.4
openai==1.56.1
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parsimonious==0.9.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prometheus-client==0.21.1
prompt-toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.3
pydantic-core==2.27.1
pygments==2.18.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pyyaml==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rlp==4.0.1
rpds-py==0.22.1
send2trash==1.8.3
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing-extensions==4.12.2
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
python-dotenv
anthropic
# Dependencias adicionales que no están en el requirements.txt original
schedule
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
contourpy==1.3.1
cycler==0.12.1
dateparser==1.2.0
fonttools==4.55.3
frozendict==2.4.6
frozenlist==1.5.0
googlesearch-python==1.2.5
html5lib==1.1
kiwisolver==1.4.8
lxml==5.3.0
matplotlib==3.10.0
multidict==6.1.0
multitasking==0.0.11
numpy==2.2.1
pandas==2.2.3
peewee==3.17.8
phi==0.6.7
pillow==11.1.0
propcache==0.2.1
pyparsing==3.2.1
python-binance==1.0.27
python-dotenv==1.0.1
pytz==2024.2
ta==0.11.0
tzdata==2024.2
tzlocal==5.2
websockets==14.1
yarl==1.18.3
yfinance==0.2.51